requirements.txt
jupyter
torch
pyro-ppl
matplotlib
scipy
torchvision
scikit-learn
tqdm